refactored GUI to gtk_gui.c, and refactored all filter files
