+++ src/filters/pixelate_filter.c
@@ -4,38 +4,35 @@
 #define MAX_COLOR 255
 #define PIXEL_SIZE 10
 
-int pixelate_filter(const char *inputFile, const char *outputFile)
+int pixelate_filter(const char *input_file, const char *output_file)
 {
-    FILE *fileIn = fopen(inputFile, "rb");
-    FILE *fileOut = fopen(outputFile, "wb+");
-    unsigned char byte[54];
+    FILE *file_in = fopen(input_file, "rb");
+    FILE *file_out = fopen(output_file, "wb+");
+    unsigned char header_info[54];
     int i, j, r, g, b;
 
-    if (fileIn == NULL || fileOut == NULL)
+    if (file_in == NULL || file_out == NULL)
     {
         printf("File does not exist.\n");
-        if (fileIn != NULL)
-            fclose(fileIn);
-        if (fileOut != NULL)
-            fclose(fileOut);
+        if (file_in != NULL)
+            fclose(file_in);
+        if (file_out != NULL)
+            fclose(file_out);
         return 1;
     }
-
     // read header info of image
     for (i = 0; i < 54; i++)
     {
-        byte[i] = getc(fileIn);
+        header_info[i] = getc(file_in);
     }
-
     // write header info to output file
-    fwrite(byte, sizeof(unsigned char), 54, fileOut);
+    fwrite(header_info, sizeof(unsigned char), 54, file_out);
     // extract attributes from image header
-    int height = *(int *)&byte[18];
-    int width = *(int *)&byte[22];
-    int size = height * width;
-
-    unsigned char buffer[size][3];
-
+    int height = *(int *)&header_info[18];
+    int width = *(int *)&header_info[22];
+    int pixels_in_image = height * width;
+    // check if image has a color table
+    unsigned char buffer[pixels_in_image][3];
     // read & write image data in chunks until end of file is reached
     for (i = 0; i < height; i += PIXEL_SIZE)
     {
@@ -44,7 +41,6 @@ int pixelate_filter(const char *inputFile, const char *outputFile)
             r = 0;
             g = 0;
             b = 0;
-
             // calculate average color of pixel block
             int count = 0;
             int x, y;
@@ -64,7 +60,6 @@ int pixelate_filter(const char *inputFile, const char *outputFile)
             r /= count;
             g /= count;
             b /= count;
-
             // set color of pixel block
             for (x = i; x < i + PIXEL_SIZE; x++)
             {
@@ -80,17 +75,15 @@ int pixelate_filter(const char *inputFile, const char *outputFile)
             }
         }
     }
-
     // write image data to output file
-    for (i = 0; i < size; i++)
+    for (i = 0; i < pixels_in_image; i++)
     {
-        putc(buffer[i][2], fileOut); // blue
-        putc(buffer[i][1], fileOut); // green
-        putc(buffer[i][0], fileOut); // red
+        putc(buffer[i][2], file_out); // blue
+        putc(buffer[i][1], file_out); // green
+        putc(buffer[i][0], file_out); // red
     }
-
-    fclose(fileIn);
-    fclose(fileOut);
+    fclose(file_in);
+    fclose(file_out);
 
     return 0;
 }
