+++ src/filters/rgb_to_gray_filter.c
@@ -1,61 +1,52 @@
 #include <stdio.h>
 
-int rgb_to_gray_filter(const char *inputFile, const char *outputFile)
+int rgb_to_gray_filter(const char *input_file, const char *output_file)
 {
-    FILE *fileIn = fopen(inputFile, "rb");
-    FILE *fileOut = fopen(outputFile, "wb+");
+    FILE *file_in = fopen(input_file, "rb");
+    FILE *file_out = fopen(output_file, "wb+");
 
-    if (fileIn == NULL || fileOut == NULL)
+    if (file_in == NULL || file_out == NULL)
     {
         printf("File does not exist.\n");
-        if (fileIn != NULL)
-            fclose(fileIn);
-        if (fileOut != NULL)
-            fclose(fileOut);
+        if (file_in != NULL)
+            fclose(file_in);
+        if (file_out != NULL)
+            fclose(file_out);
         return 1;
     }
-
-    // Read and write header info
-    unsigned char headerInfo[54];
-    fread(headerInfo, sizeof(unsigned char), 54, fileIn);
-    fwrite(headerInfo, sizeof(unsigned char), 54, fileOut);
-
-    // Extract image dimensions from header info
-    int height = *(int *)&headerInfo[18];
-    int width = *(int *)&headerInfo[22];
-    int bitDepth = *(int *)&headerInfo[28];
-    int pixelsInImage = height * width;
-
-    // Allocate Memory to Image Data
-    unsigned char(*buffer)[3] = malloc(pixelsInImage * sizeof(*buffer));
+    // read and write header info
+    unsigned char header_info[54];
+    fread(header_info, sizeof(unsigned char), 54, file_in);
+    fwrite(header_info, sizeof(unsigned char), 54, file_out);
+    // extract image dimensions from header info
+    int height = *(int *)&header_info[18];
+    int width = *(int *)&header_info[22];
+    int pixels_in_image = height * width;
+    // allocate memory to image data
+    unsigned char(*buffer)[3] = malloc(pixels_in_image * sizeof(*buffer));
     if (buffer == NULL)
     {
         printf("Memory allocation failed.\n");
-        fclose(fileIn);
-        fclose(fileOut);
+        fclose(file_in);
+        fclose(file_out);
         return 1;
     }
-
-    for (int i = 0; i < pixelsInImage; i++)
+    for (int i = 0; i < pixels_in_image; i++)
     {
-
         // read RGB color components of current pixel from input file
-        buffer[i][0] = getc(fileIn); // Red
-        buffer[i][1] = getc(fileIn); // Green
-        buffer[i][2] = getc(fileIn); // Blue
-
+        buffer[i][0] = getc(file_in); // red
+        buffer[i][1] = getc(file_in); // green
+        buffer[i][2] = getc(file_in); // blue
         // calculate grayscale value using weighted average of red and blue components
-        unsigned char grayscaleValue = (buffer[i][0] * 0.3) + (buffer[i][2] * 0.11);
-
+        unsigned char grayscale_value = (buffer[i][0] * 0.3) + (buffer[i][2] * 0.11);
         // write the calculated grayscale value to the output file for each RGB channel
-        putc(grayscaleValue, fileOut); // Red
-        putc(grayscaleValue, fileOut); // Green
-        putc(grayscaleValue, fileOut); // Blue
+        putc(grayscale_value, file_out); // red
+        putc(grayscale_value, file_out); // green
+        putc(grayscale_value, file_out); // blue
     }
-
-    // Clean up / Close
-    fclose(fileIn);
-    fclose(fileOut);
+    // clean up memory
+    fclose(file_in);
+    fclose(file_out);
 
     return 0;
 }
